refactor(notifier): split markdown→Notion conversion into markdown.py (#217)
* refactor(notifier): split markdown→Notion conversion into markdown.py

Move _markdown_to_blocks and all inline/block helpers out of notion.py
into a dedicated src/notifier/markdown.py module. notion.py now only
contains Notion API client logic. Public function renamed to
markdown_to_notion_blocks; _markdown_to_blocks alias kept for tests.


* fix(markdown): strip only bold delimiters from headings; escape fullwidth colon

- re.sub pattern changed from r'\*+' to r'^\*{2}|\*{2}$' so lone
  asterisks (e.g. '2 * 3', product names) are preserved in headings
- _split_label_colon: encode fullwidth colon as '\uFF1A' to avoid
  Ruff flagging the literal Unicode character


---------
